+++ rollup.config.mjs
@@ -1,7 +1,6 @@
 import { babel } from "@rollup/plugin-babel";
 import dts from "rollup-plugin-dts";
 import vue from "rollup-plugin-vue";
-import copy from "rollup-plugin-copy";
 import postcss from "rollup-plugin-postcss";
 
 export default [
@@ -25,39 +24,12 @@ export default [
         babelHelpers: 'runtime',
         exclude: ["./node_modules/**", "./qa/**"],
       }),
-      copy({
-        targets: [
-          { src: 'src/assets/images/*', dest: 'dist/assets/images' }
-        ],
-      }),
       vue({
         preprocessStyles: true,
       }),
       postcss({
         minimize: true,
       }),
-      {
-        transform(code) {
-          const importMetaRegExp = /require\('url'\).pathToFileURL\(__filename\).toString\(\)/g;
-          const assetsRegExp = /\.\.\/assets\/images/g;
-          const importMetaMatches = Array.from(code.matchAll(importMetaRegExp));
-          const assetsMatches = Array.from(code.matchAll(assetsRegExp));
-          let newCode = code;
-
-          for (const [all] of importMetaMatches) {
-            newCode = newCode.replace(all, "import.meta.url");
-          }
-
-          for (const [all] of assetsMatches) {
-            newCode = newCode.replace(all, "./assets/images");
-          }
-
-          return {
-            code: newCode,
-            map: {mappings: ""},
-          };
-        }
-      }
     ]
   },
   {